StructureUpdate
Updated the structure of the files to separate the operations into separate functions. This includes moving the test into a separate file that calls the optimization framework. A second optimization function has been constructed that changes how the Truth Model is queried.
