--- current_index.txt
@@ -1,1 +1,0 @@
-23
