+++ data/init_data_test.csv
@@ -1,0 +1,5 @@
+3.140061087334178325e-01,1.663579075733306878e-01,3.970868410602836907e-01,8.769434871064207826e-02
+3.695977414280995355e-01,6.340219370356948714e-01,6.642598295683512788e-01,2.632647513453965082e-01
+6.194423112371214568e-01,5.569468858105313114e-01,3.454721955313034831e-01,5.132789334514223922e-01
+7.517669566786879809e-01,5.763508570364235251e-01,1.575250626644729701e-01,6.055257125735313828e-02
+3.338659699536892234e-01,6.962811838178339929e-02,8.166483815492084908e-01,7.924372352304832612e-01
